+++ runner-core/src/main/java/org/apache/apisix/plugin/runner/server/ApplicationRunner.java
@@ -1,5 +1,6 @@
 package org.apache.apisix.plugin.runner.server;
 
+import com.google.common.cache.Cache;
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelInitializer;
@@ -10,10 +11,13 @@ import io.netty.channel.unix.DomainSocketAddress;
 import io.netty.channel.unix.DomainSocketChannel;
 import io.netty.handler.logging.LoggingHandler;
 import lombok.RequiredArgsConstructor;
+import org.apache.apisix.plugin.runner.A6Conf;
+import org.apache.apisix.plugin.runner.filter.PluginFilter;
 import org.apache.apisix.plugin.runner.handler.BinaryProtocolDecoder;
 import org.apache.apisix.plugin.runner.handler.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
@@ -22,6 +26,10 @@ import org.springframework.stereotype.Component;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Component
 @RequiredArgsConstructor
@@ -32,14 +40,27 @@ public class ApplicationRunner implements CommandLineRunner {
     @Value("${socket.file}")
     private String socketFile;
 
-    private PrepareConfHandler prepareConfHandler;
+    private Cache<Long, A6Conf> cache;
 
-    private HTTPReqCallHandler hTTPReqCallHandler;
+    private ObjectProvider<PluginFilter> beanProvider;
 
     @Autowired
-    public ApplicationRunner(PrepareConfHandler prepareConfHandler, HTTPReqCallHandler hTTPReqCallHandler) {
-        this.prepareConfHandler = prepareConfHandler;
-        this.hTTPReqCallHandler = hTTPReqCallHandler;
+    public ApplicationRunner(Cache<Long, A6Conf> cache, ObjectProvider<PluginFilter> beanProvider) {
+        this.cache = cache;
+        this.beanProvider = beanProvider;
+    }
+
+    public PrepareConfHandler createConfigReqHandler(Cache<Long, A6Conf> cache, ObjectProvider<PluginFilter> beanProvider) {
+        List<PluginFilter> pluginFilterList = beanProvider.orderedStream().collect(Collectors.toList());
+        Map<String, PluginFilter> filterMap = new HashMap<>();
+        for (PluginFilter filter : pluginFilterList) {
+            filterMap.put(filter.name(), filter);
+        }
+        return new PrepareConfHandler(cache, filterMap);
+    }
+
+    public HTTPReqCallHandler createA6HttpHandler(Cache<Long, A6Conf> cache) {
+        return new HTTPReqCallHandler(cache);
     }
 
     public void start(String path) throws Exception {
@@ -63,13 +84,12 @@ public class ApplicationRunner implements CommandLineRunner {
                 .childHandler(new ChannelInitializer<DomainSocketChannel>() {
                     @Override
                     protected void initChannel(DomainSocketChannel channel) {
-                        channel.pipeline().addFirst(new LoggingHandler());
                         channel.pipeline().addFirst("logger", new LoggingHandler())
                                 .addAfter("logger","payloadEncoder",new PayloadEncoder())
                                 .addAfter("payloadEncoder", "delayedDecoder", new BinaryProtocolDecoder())
                                 .addLast("payloadDecoder", new PayloadDecoder())
-                                .addAfter("payloadDecoder", "prepareConfHandler", prepareConfHandler)
-                                .addAfter("prepareConfHandler", "hTTPReqCallHandler", hTTPReqCallHandler);
+                                .addAfter("payloadDecoder", "prepareConfHandler", createConfigReqHandler(cache, beanProvider))
+                                .addAfter("prepareConfHandler", "hTTPReqCallHandler", createA6HttpHandler(cache));
 
                     }
                 });
